requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.5
joblib==1.3.2
torch==2.0.1
transformers==4.35.2
gunicorn==21.2.0